refactor: move LedgerAccount struct to value.rs
